encrypted_traffic_ids/requirements.txt
# Deep Learning Frameworks
torch==2.1.0
torchvision==0.16.0
torch-geometric==2.4.0
transformers==4.35.0

# Scientific Computing
numpy==1.26.0
scipy==1.11.3
pandas==2.1.0

# Machine Learning
scikit-learn==1.3.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0

# Explainability
shap==0.43.0

# Federated Learning & Privacy
opacus==1.4.0
cryptography==41.0.5
phe==1.5.0

# Network Analysis
scapy==2.5.0
dpkt==1.9.8

# Utilities
tqdm==4.66.1
joblib==1.3.2
tensorboard==2.15.0
pyyaml==6.0.1
h5py==3.10.0

# Model Optimization
onnx==1.15.0
onnxruntime==1.16.0

# Metrics and Evaluation
torchmetrics==1.2.0
